# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Recommendation Engine API
  description: |
    A comprehensive recommendation engine API that provides personalized content recommendations
    using multiple algorithms including semantic search, collaborative filtering, and graph-based approaches.
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.recommendations.example.com/api/v1
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /content/jobs/{jobId}:
    get:
      summary: Get content processing job status
      description: Check the status of a content ingestion job
      operationId: getJobStatus
      tags:
        - Content Management
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Job ID returned from content ingestion
      responses:
        '200':
          description: Job status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /recommendations/{userId}:
    get:
      summary: Get personalized recommendations
      description: Retrieve personalized content recommendations for a specific user
      operationId: getRecommendations
      tags:
        - Recommendations
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID to get recommendations for
        - name: count
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of recommendations to return
        - name: context
          in: query
          schema:
            type: string
            enum: [home, search, product_page, category, checkout]
            default: home
          description: Context where recommendations will be displayed
        - name: explain
          in: query
          schema:
            type: boolean
            default: false
          description: Include explanation for each recommendation
        - name: categories
          in: query
          schema:
            type: array
            items:
              type: string
          style: form
          explode: false
          description: Filter recommendations by categories
        - name: exclude
          in: query
          schema:
            type: array
            items:
              type: string
          style: form
          explode: false
          description: Item IDs to exclude from recommendations
      responses:
        '200':
          description: Recommendations retrieved successfully
          headers:
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/RateLimitError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
...
    Recommendation:
      type: object
      properties:
        itemId:
          type: string
          description: ID of the recommended content item
        score:
          type: number
          minimum: 0
          maximum: 1
          description: Recommendation score (0-1)
        algorithm:
          type: string
          enum: [semantic_search, collaborative_filtering, pagerank, popularity, hybrid]
          description: Algorithm used to generate this recommendation
        explanation:
          $ref: '#/components/schemas/Explanation'
        confidence:
          type: number
          minimum: 0
          maximum: 1
          description: Confidence score for this recommendation
        position:
          type: integer
          minimum: 1
          description: Position in the recommendation list
        metadata:
          type: object
          additionalProperties: true
          description: Additional metadata about the recommendation
      required:
        - itemId
        - score
        - algorithm
        - confidence
        - position
    Explanation:
      type: object
      properties:
        type:
          type: string
          enum: [content_based, collaborative, graph_based, popularity_based]
          description: Type of explanation
        message:
          type: string
          description: Human-readable explanation
          example: "Because you liked 'Wireless Mouse' and both are in electronics category"
        evidence:
          type: object
          properties:
            similarItems:
              type: array
              items:
                type: string
              description: Items that contributed to this recommendation
            sharedUsers:
              type: integer
              description: Number of users with similar preferences
            categories:
              type: array
              items:
                type: string
              description: Shared categories
            confidence:
              type: number
              minimum: 0
              maximum: 1
        details:
          type: object
          additionalProperties: true
          description: Additional explanation details
      required:
        - type
        - message
...
    JobStatus:
      type: object
      properties:
        jobId:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, processing, completed, failed]
        progress:
          type: number
          minimum: 0
          maximum: 100
          description: Processing progress percentage
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        error:
          type: string
          description: Error message if job failed
        results:
          type: object
          properties:
            embeddingsGenerated:
              type: boolean
            metadataStored:
              type: boolean
            indexUpdated:
              type: boolean
    RecommendationResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        recommendations:
          type: array
          items:
            $ref: '#/components/schemas/Recommendation'
        metadata:
          type: object
          properties:
            totalAvailable:
              type: integer
            algorithmsUsed:
              type: array
              items:
                type: string
            generatedAt:
              type: string
              format: date-time
            cacheHit:
              type: boolean
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
        hasMore:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              additionalProperties: true
              description: Additional error details
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
              format: uuid
              description: Unique request identifier for debugging
          required:
            - code
            - message
            - timestamp
        fallback:
          type: object
          properties:
            used:
              type: boolean
              description: Whether fallback data was used
            strategy:
              type: string
              description: Fallback strategy employed
            confidence:
              type: number
              minimum: 0
              maximum: 1
              description: Confidence in fallback data
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            invalid_content_type:
              summary: Invalid content type
              value:
                error:
                  code: "INVALID_CONTENT_TYPE"
                  message: "Content type must be one of: product, video, article, course, book"
                  details:
                    field: "type"
                    provided: "invalid_type"
                    allowed: ["product", "video", "article", "course", "book"]
                  timestamp: "2024-01-15T10:30:00Z"
                  requestId: "req-123"
            missing_required_field:
              summary: Missing required field
              value:
                error:
                  code: "MISSING_REQUIRED_FIELD"
                  message: "Required field 'title' is missing"
                  details:
                    field: "title"
                    required: true
                  timestamp: "2024-01-15T10:30:00Z"
                  requestId: "req-124"
    UnauthorizedError:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Valid JWT token required"
              details:
                reason: "token_expired"
              timestamp: "2024-01-15T10:30:00Z"
              requestId: "req-125"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            user_not_found:
              summary: User not found
              value:
                error:
                  code: "USER_NOT_FOUND"
                  message: "User with ID 'user-123' not found"
                  details:
                    userId: "user-123"
                  timestamp: "2024-01-15T10:30:00Z"
                  requestId: "req-126"
                fallback:
                  used: true
                  strategy: "popular_recommendations"
                  confidence: 0.3
    RateLimitError:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Requests remaining in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when limit resets
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "RATE_LIMIT_EXCEEDED"
              message: "Rate limit of 1000 requests per hour exceeded"
              details:
                limit: 1000
                window: "1h"
                retryAfter: 3600
              timestamp: "2024-01-15T10:30:00Z"
              requestId: "req-127"
...